requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.2.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.1.0
idna==2.10
psycopg2==2.9.1
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0